requirements.txt
pandas~=1.4.4
streamlit~=1.21.0
requests~=2.28.1
matplotlib~=3.5.3
numpy~=1.23.3
extra-streamlit-components~=0.1.56
python-dateutil~=2.8.2